refactor: use base plugin for web and node
