+++ lib/node/ReadFileCompileWasmTemplatePlugin.js
@@ -4,7 +4,7 @@
 */
 "use strict";
 
-const FetchCompileWasmMainTemplatePlugin = require("../web/FetchCompileWasmMainTemplatePlugin");
+const BaseWasmMainTemplatePlugin = require("../BaseWasmMainTemplatePlugin");
 const WasmModuleTemplatePlugin = require("../wasm/WasmModuleTemplatePlugin");
 
 class ReadFileCompileWasmTemplatePlugin {
@@ -12,9 +12,7 @@ class ReadFileCompileWasmTemplatePlugin {
 		compiler.hooks.thisCompilation.tap(
 			"ReadFileCompileWasmTemplatePlugin",
 			compilation => {
-				new FetchCompileWasmMainTemplatePlugin().apply(
-					compilation.mainTemplate
-				);
+				new BaseWasmMainTemplatePlugin().applyNode(compilation.mainTemplate);
 				new WasmModuleTemplatePlugin().apply(
 					compilation.moduleTemplates.javascript
 				);
